[#3935] Add ycqlsh and ycqlsh.py to cqlsh repo
Summary:
See https://github.com/yugabyte/yugabyte-db/issues/3935

The plan is to copy these files into the bin directory instead of the cqlsh/cqlsh.py files.

I have attempted to patch some references to cqlsh but my goal is not to comprehensively eliminate all traces of cqlsh. For ex: I think it still uses cqlshrc and I don't change it to .cqlshrc. The main goal of the rename is to emphasize the difference in behavior between our CQL implementation and the vanilla Cassandra ones and I think having a different binary name already achieves that.

Test Plan: Run bin/ycqlsh, verify it works.

Reviewers: bogdan, jason, wesley

Reviewed By: wesley

Subscribers: ybase

Differential Revision: https://phabricator.dev.yugabyte.com/D8240